feat(front): check overlapping
