+++ front/package.json
@@ -17,6 +17,7 @@
     "framer-motion": "^10.16.1",
     "jotai": "^2.4.0",
     "jotai-tanstack-query": "^0.7.1",
+    "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "ts-pattern": "^5.0.5"
@@ -24,6 +25,7 @@
   "devDependencies": {
     "@pandacss/dev": "^0.13.0",
     "@tanstack/react-query-devtools": "^4.33.0",
+    "@types/lodash": "^4.14.197",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
