--- front/src/utils/dom_util.ts
@@ -1,3 +1,0 @@
-export const isIntersect = (r1: DOMRect, r2: DOMRect) => {
-  return !(r2.left > r1.right || r2.right < r1.left || r2.top > r1.bottom || r2.bottom < r1.top);
-};
